requirements.txt
google-generativeai
langchain
langchain-google-genai
sentence-transformers
chromadb
rank_bm25
streamlit
pandas
python-dotenv
faiss-cpu
langchain-groq
fastapi
uvicorn